bug 740, change BYTE to INTEGER*1

+++ fortran/ccplib.f
@@ -94,10 +94,10 @@ C_END_CCFILL
 C
 C     .. Scalar Arguments ..
       INTEGER NTIMES
-      BYTE SCAL
+      INTEGER*1 SCAL
 C     ..
 C     .. Array Arguments ..
-      BYTE ARR1(*)
+      INTEGER*1 ARR1(*)
 C     ..
 C     .. Local Scalars ..
       INTEGER N
@@ -253,8 +253,8 @@ C SPECIFICATION STATEMENTS
 C ------------------------
 C
       INTEGER IA(*)
-      BYTE IBYT(*)
-      BYTE JBYT(4)
+      INTEGER*1 IBYT(*)
+      INTEGER*1 JBYT(4)
       EQUIVALENCE (JA,JBYT(1))
       LOGICAL CALLED, LITEND
       INTEGER IND
@@ -348,9 +348,9 @@ C
 C====== Specification statements
 C
       INTEGER IA(*)
-      BYTE IB(*)
+      INTEGER*1 IB(*)
       INTEGER*2 J2(2)
-      BYTE JBYT(4)
+      INTEGER*1 JBYT(4)
       EQUIVALENCE (JA,J2(1),JBYT(1))
       LOGICAL CALLED, LITEND
       EXTERNAL LITEND
@@ -668,8 +668,8 @@ C
 C SPECIFICATION STATEMENTS
 C ------------------------
 C
-      BYTE IA(*)
-      BYTE JBYT(4)
+      INTEGER*1 IA(*)
+      INTEGER*1 JBYT(4)
       EQUIVALENCE (JA,JBYT(1))
       LOGICAL CALLED, LITEND
       EXTERNAL LITEND
@@ -799,8 +799,8 @@ C SPECIFICATION STATEMENTS
 C ------------------------
 C
       INTEGER IA(*)
-      BYTE IBYT(*)
-      BYTE JBYT(4)
+      INTEGER*1 IBYT(*)
+      INTEGER*1 JBYT(4)
       EQUIVALENCE (JA,JBYT(1))
       LOGICAL CALLED, LITEND
       EXTERNAL LITEND
@@ -990,7 +990,7 @@ C     .. Scalar Arguments ..
       INTEGER I1,I2,NTOMOV
 C     ..
 C     .. Array Arguments ..
-      BYTE ARR1(*),ARR2(*)
+      INTEGER*1 ARR1(*),ARR2(*)
 C     ..
 C     .. Local Scalars ..
       INTEGER I,J,N
@@ -1602,8 +1602,8 @@ C
 C SPECIFICATION STATEMENTS
 C ------------------------
 C
-      BYTE IA(*)
-      BYTE JBYT(4)
+      INTEGER*1 IA(*)
+      INTEGER*1 JBYT(4)
       EQUIVALENCE (JA,JBYT(1))
       EQUIVALENCE (JA,JBYT(1))
       LOGICAL CALLED, LITEND
@@ -1705,7 +1705,7 @@ C     .. Local Scalars ..
       INTEGER IA,NB,NIH,NW
 C     ..
 C     .. Local Arrays ..
-      BYTE IBYT(4),JBYT(4)
+      INTEGER*1 IBYT(4),JBYT(4)
       INTEGER*2 JHALF(2)
 C     ..
 C     .. Intrinsic Functions ..
@@ -1817,7 +1817,7 @@ C_END_CCPZBI
 C
 C  Arguments ......
       INTEGER NUM
-      BYTE ARR1(*)
+      INTEGER*1 ARR1(*)
 C
       INTEGER J
 C
@@ -1948,7 +1948,7 @@ C    IDUM (D)   DUMMY
 C_END_LITEND
 C
          INTEGER I, IDUM
-         BYTE B(4)
+         INTEGER*1 B(4)
          EQUIVALENCE (I,B(1))
 C
 C---- Initialise B
